+++ pac4j-javaee/src/main/java/org/pac4j/jee/context/JEEContext.java
@@ -47,6 +47,8 @@ public class JEEContext implements WebContext {
 
     @Override
     public Optional<String> getRequestParameter(final String name) {
+        // force reading request content not to lose it as request.getParameter may consume it
+        getRequestContent();
         return Optional.ofNullable(this.request.getParameter(name));
     }