+++ pac4j-jakartaee/src/main/java/org/pac4j/jee/context/JEEContext.java
@@ -177,7 +177,7 @@ public class JEEContext implements WebContext {
         if (queryString == null) {
             return requestURL.toString();
         }
-        return requestURL.append('?').append(queryString).toString();
+        return requestURL.toString() + '?' + queryString;
     }
 
     /** {@inheritDoc} */