+++ pac4j-javaee/src/main/java/org/pac4j/jee/context/JEEContext.java
@@ -159,7 +159,7 @@ public class JEEContext implements WebContext {
         if (queryString == null) {
             return requestURL.toString();
         }
-        return requestURL.append('?').append(queryString).toString();
+        return requestURL.toString() + '?' + queryString;
     }
 
     @Override