+++ documentation/docs/release-notes.md
@@ -7,6 +7,7 @@ title: Release notes&#58;
 
 **v6.1.2**:
 - Use the configured scope in OpenID Connect authenticator
+- Fix the `getFullRequestURL` method
 
 **v6.1.1**:
 - Protect the `getRequestAttribute` method for Jetty 12.0.8+
@@ -86,6 +87,7 @@ title: Release notes&#58;
 
 **v5.7.8**:
 - Fix bug for HTML values in POST forms
+- Fix the `getFullRequestURL` method
 
 **v5.7.7**:
 - Security fix: cannot accept empty OIDC credentials